requirements.txt
django==4.2.6
cloudinary==1.36.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
pika==1.3.2
pillow==10.1.0
django-cloudinary-storage==0.3.0